Added docker-compose file.
For postgres and redis we specify just the image used from dockerhub.

For the server, client and worker we specify the build and volumes.
For the server we also set up the environment variables in the
container at run time which are needed for successful connection to
both redis and postgres.